Waypoints got deleted (again...)
I think the merge from the code laptop deleted all of the waypoints, so I basically redid all of the waypoints and created a 'RECOVERY' path that contains all of the waypoints, to prevent this from happening again. As when a waypoint is being actively used in a path, it is not lost.

I also created a google sheet to store all of the waypoint data if they somehow get deleted again.